some more structuring of the project
