+++ code/src/App.js
@@ -6,7 +6,6 @@ import Form from './components/Form'
 
 export const App = () => {
 
-
   return (
     <>
       <div>
