--- code/src/components/Button.js
@@ -1,26 +1,0 @@
-// import React, { useState } from 'react'
-
-// const Button = () => {
-
-//     const onQuestionChange = ({ newQuestion, setNewQuestion }) => {
-//         console.log(newQuestion)
-//         setNewQuestion(newQuestion +1)
-//     }
-
-//     return (
-//         <>
-//             <button
-//                 type="button" 
-//                 onClick={onQuestionChange}>
-//                 Let's go!
-//             </button>
-            
-//             {/* {newQuestion > 0 && <Form/>} */}
-                
-//         </>
-//     )
-// }
-
-// export default Button;
-
-
