+++ code/src/components/ButtonNext.js
@@ -1,0 +1,24 @@
+import React, { useState } from 'react'
+
+const ButtonNext = ({ questionIndex, setQuestionIndex }) => {
+
+    const onQuestionIndexChange = () => {
+        console.log(questionIndex)
+        setQuestionIndex(questionIndex +1)
+    }
+
+    return (
+        <>
+            <button 
+                type="button"
+                onClick={onQuestionIndexChange}>
+                Next
+            </button>
+                
+        </>
+    )
+}
+
+export default ButtonNext;
+
+
