+++ code/src/components/QuestionThree.js
@@ -1,7 +1,7 @@
-import React from 'react';
+import React from 'react'
 // import RadioOption from './RadioOption'
 
-// const flowers = ['rose', 'tulip', 'orchid']
+
 
 const QuestionThree = ({ flower, setFlower, item }) => {
 
@@ -11,11 +11,6 @@ const QuestionThree = ({ flower, setFlower, item }) => {
 
     return (
         <>
-            {/* <div>
-                <h2>Third Question</h2>
-                <p>What's your favorite flower?</p>
-             </div> */}
-
             <label key={item} htmlFor={item}>{item}</label>
                 <input 
                     name="flower" 
@@ -25,38 +20,6 @@ const QuestionThree = ({ flower, setFlower, item }) => {
                     onChange={onFlowerChange}
                     checked={flower === item }  
                 />     
-  
-            {/* <div>
-                <label htmlFor="Rose">Rose</label>
-                <input 
-                    name="flower" 
-                    id="rose" 
-                    type="radio" 
-                    value="rose"
-                    onChange={onFlowerChange}
-                    checked={flower === "rose"}  
-                />
-
-                <label htmlFor="Rose">Tulip</label>
-                <input 
-                    name="flower" 
-                    id="tulip" 
-                    type="radio" 
-                    value="tulip" 
-                    onChange={onFlowerChange} 
-                    checked={flower === "tulip"}  
-                />
-
-                <label htmlFor="Orchid">Orchid</label>
-                <input 
-                    name="flower" 
-                    id="orchid" 
-                    type="radio" 
-                    value="orchid" 
-                    onChange={onFlowerChange} 
-                    checked={flower === "orchid"} 
-                />
-            </div>  */}
         </>
     )
