+++ code/src/components/QuestionTwo.js
@@ -10,10 +10,6 @@ const QuestionTwo = ({ food, setFood }) => {
     return (
         <>
             <div>
-                <h2>Second Question</h2>
-                <p>What's your favorite food?</p>
-             </div>
-            <div>
                 <label htmlFor="food"></label>
                 <select 
                     id="food"
@@ -26,7 +22,6 @@ const QuestionTwo = ({ food, setFood }) => {
                     <option>Soup</option>
                     <option>Sallad</option>
                 </select>
-                <p>Chosen food: {food}</p>
             </div>
         </>
     )
