+++ code/src/components/Summary.js
@@ -6,7 +6,7 @@ const Summary = ({ name, food, flower }) => {
     return (
         <>
             <div>
-                <h1>Hurra, en summary</h1>
+                <h1>Thank you!</h1>
                 <h3>Your answers: Your name: {name}. You're favorite food: {food} Your favorite flower is {flower}.</h3>
             </div>
         </>
